+++ src/main/kotlin/io/openfuture/chain/property/ConsensusProperties.kt
@@ -34,6 +34,30 @@ class ConsensusProperties(
      * Genesis address
      */
     @field:NotNull
-    var genesisAddress: String? = null
+    var genesisAddress: String? = null,
+
+    /**
+     * Reward for block
+     */
+    @field:NotNull
+    var rewardBlock: Double? = null,
+
+    /**
+     * Fee for coinbase transaction
+     */
+    @field:NotNull
+    var feeCoinBaseTx: Double? = null,
+
+    /**
+     * Fee for transfer transaction
+     */
+    @field:NotNull
+    var feeTransferTx: Double? = null,
+
+    /**
+     * Fee for vote transaction
+     */
+    @field:NotNull
+    var feeVoteTx: Double? = null
 
 )