+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Sign transactions
 - Create and save CoinBaseTransaction
 - Validation of CoinBaseTransaction
+- Constants fees and reward for transactions/block
 
 ## [0.3.0] - 2018-07-13
 ### Added